Split ObservationStoreTest
ObservationStoreTest was getting huge. It was already organized into nested test
classes, so move all the nested classes up to the top level.

No changes to the actual test logic here, just rename and move refactorings.
